flavors wip
