--- packages/cycle-scripts/scripts/index.ejs
@@ -1,10 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-    <meta charset="utf-8">
-    <title>INSERT APP NAME HERE</title>
-</head>
-<body>
-    <div id="app" />
-</body>
-</html>
