+++ packages/cycle-scripts/scripts/init.js
@@ -4,49 +4,11 @@ const fs = require('fs-extra')
 const path = require('path')
 const chalk = require('chalk')
 const spawn = require('cross-spawn')
-const initQuestions = require('./initQuestions')
+const initQuestions = require('./utils/initQuestions')
 // Ask the user for which language and which stream library he want to use
-const dependencies = {
-  basics: [
-    '@cycle/dom@17.1.0'
-  ],
-  language: {
-    'JavaScript': [],
-    'TypeScript': []
-  },
-  streamLib: {
-    xstream: [
-      '@cycle/run@3.1.0',
-      'xstream@10.5.0'
-    ],
-    rxjs: [
-      '@cycle/rxjs-run@7.0.0',
-      'rxjs@5.3.0'
-    ],
-    most: [
-      '@cycle/most-run@7.1.0',
-      'most@1.2.2'
-    ]
-  }
-}
-
-const replacements = {
-  xstream: {
-    run: '@cycle/run',
-    import: 'import xs from \'xstream\'',
-    stream: 'xs'
-  },
-  rxjs: {
-    run: '@cycle/rxjs-run',
-    import: 'import Rx from \'rxjs/Rx\'',
-    stream: 'Rx.Observable'
-  },
-  most: {
-    run: '@cycle/most-run',
-    import: 'import * as most from \'most\'',
-    stream: 'most'
-  }
-}
+const flavors = require('../configs/flavors')
+const dependencies = flavors.dependencies
+const replacements = flavors.replacements
 
 // function patchGitignore (appPath) {
 //   // Rename gitignore after the fact to prevent npm from renaming it to .npmignore
