+++ packages/cycle-scripts/scripts/utils/initQuestions.js
